requirements.txt
Flask-API==2.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
Flask-Script==2.0.6
Flask==1.1.2
SQLAlchemy==1.3.19
flask-marshmallow==0.14.0
flask-shell-ipython==0.4.1
marshmallow-sqlalchemy==0.23.1
psycopg2==2.8.6
pytest-flask-sqlalchemy==1.0.2
pytest-flask==1.0.0
pytest==6.1.1
python-dotenv==0.14.0
